+++ csvkit/cli.py
@@ -141,7 +141,7 @@ class CSVKitUtility(object):
                                         help='Delimiting character of the input CSV file.')
         if 't' not in self.override_flags:
             self.argparser.add_argument('-t', '--tabs', dest='tabs', action='store_true',
-                                        help='Specifies that the input CSV file is delimited with tabs. Overrides "-d".')
+                                        help='Specify that the input CSV file is delimited with tabs. Overrides "-d".')
         if 'q' not in self.override_flags:
             self.argparser.add_argument('-q', '--quotechar', dest='quotechar',
                                         help='Character used to quote strings in the input CSV file.')
@@ -159,13 +159,25 @@ class CSVKitUtility(object):
                                         help='Maximum length of a single field in the input CSV file.')
         if 'e' not in self.override_flags:
             self.argparser.add_argument('-e', '--encoding', dest='encoding', default='utf-8',
-                                        help='Specify the encoding the input CSV file.')
+                                        help='Specify the encoding of the input CSV file.')
+        if 'L' not in self.override_flags:
+            self.argparser.add_argument('-L', '--locale', dest='locale', default='en_US',
+                                        help='Specify the locale (en_US) of any formatted numbers.')
         if 'S' not in self.override_flags:
             self.argparser.add_argument('-S', '--skipinitialspace', dest='skipinitialspace', action='store_true',
                                         help='Ignore whitespace immediately following the delimiter.')
+        if 'date-format' not in self.override_flags:
+            self.argparser.add_argument('--date-format', dest='date_format',
+                                        help='Specify a strptime date format string like "%%m/%%d/%%Y".')
+        if 'datetime-format' not in self.override_flags:
+            self.argparser.add_argument('--datetime-format', dest='datetime_format',
+                                        help='Specify a strptime datetime format string like "%%m/%%d/%%Y %%I:%%M %%p".')
         if 'H' not in self.override_flags:
             self.argparser.add_argument('-H', '--no-header-row', dest='no_header_row', action='store_true',
-                                        help='Specifies that the input CSV file has no header row. Will create default headers (A,B,C,...).')
+                                        help='Specify that the input CSV file has no header row. Will create default headers (a,b,c,...).')
+        if 'K' not in self.override_flags:
+            self.argparser.add_argument('-K', '--skip-lines', dest='skip_lines', type=int, default=0,
+                                        help='Specify the number of initial lines to skip (e.g. comments, copyright notices, empty rows).')
         if 'v' not in self.override_flags:
             self.argparser.add_argument('-v', '--verbose', dest='verbose', action='store_true',
                                         help='Print detailed tracebacks when errors occur.')
@@ -179,6 +191,9 @@ class CSVKitUtility(object):
         if 'zero' not in self.override_flags:
             self.argparser.add_argument('--zero', dest='zero_based', action='store_true',
                                         help='When interpreting or displaying column numbers, use zero-based numbering instead of the default 1-based numbering.')
+
+        self.argparser.add_argument('-V', '--version', action='version', version='%(prog)s 1.0.2',
+                                        help='Display version information and exit.')
 
     def _open_input_file(self, path):
         """
@@ -230,6 +245,9 @@ class CSVKitUtility(object):
         if six.PY2 and self.args.encoding:
             kwargs['encoding'] = self.args.encoding
 
+        if getattr(self.args, 'no_header_row', None):
+            kwargs['header'] = not self.args.no_header_row
+
         return kwargs
 
     def _extract_csv_writer_kwargs(self):
@@ -238,7 +256,7 @@ class CSVKitUtility(object):
         """
         kwargs = {}
 
-        if 'l' not in self.override_flags and self.args.line_numbers:
+        if getattr(self.args, 'line_numbers', None):
             kwargs['line_numbers'] = True
 
         return kwargs
@@ -274,10 +292,10 @@ class CSVKitUtility(object):
         else:
             return agate.TypeTester(types=[
                 agate.Boolean(),
-                agate.Number(),
+                agate.Number(locale=self.args.locale),
                 agate.TimeDelta(),
-                agate.Date(),
-                agate.DateTime(),
+                agate.Date(date_format=self.args.date_format),
+                agate.DateTime(datetime_format=self.args.datetime_format),
                 text_type
             ])
 
@@ -287,8 +305,18 @@ class CSVKitUtility(object):
         else:
             return 1
 
+    def skip_lines(self):
+        if isinstance(self.args.skip_lines, int):
+            while self.args.skip_lines > 0:
+                self.input_file.readline()
+                self.args.skip_lines -= 1
+        else:
+            raise ValueError('skip_lines argument must be an int')
+
+        return self.input_file
+
     def get_rows_and_column_names_and_column_ids(self, **kwargs):
-        rows = agate.csv.reader(self.input_file, **kwargs)
+        rows = agate.csv.reader(self.skip_lines(), **kwargs)
 
         if self.args.no_header_row:
             # Peek at a row to get the number of columns.