+++ csvkit/utilities/csvgrep.py
@@ -11,6 +11,7 @@ from csvkit.grep import FilteringCSVReader
 
 class CSVGrep(CSVKitUtility):
     description = 'Search CSV files. Like the Unix "grep" command, but for tabular data.'
+    override_flags = ['L', 'date-format', 'datetime-format']
 
     def add_arguments(self):
         self.argparser.add_argument('-n', '--names', dest='names_only', action='store_true',
@@ -39,8 +40,9 @@ class CSVGrep(CSVKitUtility):
 
         reader_kwargs = self.reader_kwargs
         writer_kwargs = self.writer_kwargs
+        # Move the line_numbers option from the writer to the reader.
         if writer_kwargs.pop('line_numbers', False):
-            reader_kwargs = {'line_numbers': True}
+            reader_kwargs['line_numbers'] = True
 
         rows, column_names, column_ids = self.get_rows_and_column_names_and_column_ids(**reader_kwargs)